Litz ebourne template level (#191)
