--- include/GMGPolar/level_interpolation.inl
@@ -1,69 +1,0 @@
-template <concepts::DomainGeometry DomainGeometry, concepts::DensityProfileCoefficients DensityProfileCoefficients>
-void GMGPolar<DomainGeometry, DensityProfileCoefficients>::prolongation(int current_level, Vector<double> result,
-                                                                        ConstVector<double> x) const
-{
-    assert(current_level < number_of_levels_ && 1 <= current_level);
-    if (!interpolation_)
-        throw std::runtime_error("Interpolation not initialized.");
-
-    interpolation_->applyProlongation(levels_[current_level].grid(), levels_[current_level - 1].grid(), result, x);
-}
-
-template <concepts::DomainGeometry DomainGeometry, concepts::DensityProfileCoefficients DensityProfileCoefficients>
-void GMGPolar<DomainGeometry, DensityProfileCoefficients>::restriction(int current_level, Vector<double> result,
-                                                                       ConstVector<double> x) const
-{
-    assert(current_level < number_of_levels_ - 1 && 0 <= current_level);
-    if (!interpolation_)
-        throw std::runtime_error("Interpolation not initialized.");
-
-    interpolation_->applyRestriction(levels_[current_level].grid(), levels_[current_level + 1].grid(), result, x);
-}
-
-template <concepts::DomainGeometry DomainGeometry, concepts::DensityProfileCoefficients DensityProfileCoefficients>
-void GMGPolar<DomainGeometry, DensityProfileCoefficients>::injection(int current_level, Vector<double> result,
-                                                                     ConstVector<double> x) const
-{
-    assert(current_level < number_of_levels_ - 1 && 0 <= current_level);
-    if (!interpolation_)
-        throw std::runtime_error("Interpolation not initialized.");
-
-    interpolation_->applyInjection(levels_[current_level].grid(), levels_[current_level + 1].grid(), result, x);
-}
-
-template <concepts::DomainGeometry DomainGeometry, concepts::DensityProfileCoefficients DensityProfileCoefficients>
-void GMGPolar<DomainGeometry, DensityProfileCoefficients>::extrapolatedProlongation(int current_level,
-                                                                                    Vector<double> result,
-                                                                                    ConstVector<double> x) const
-{
-    assert(current_level < number_of_levels_ && 1 <= current_level);
-    if (!interpolation_)
-        throw std::runtime_error("Interpolation not initialized.");
-
-    interpolation_->applyExtrapolatedProlongation(levels_[current_level].grid(), levels_[current_level - 1].grid(),
-                                                  result, x);
-}
-
-template <concepts::DomainGeometry DomainGeometry, concepts::DensityProfileCoefficients DensityProfileCoefficients>
-void GMGPolar<DomainGeometry, DensityProfileCoefficients>::extrapolatedRestriction(int current_level,
-                                                                                   Vector<double> result,
-                                                                                   ConstVector<double> x) const
-{
-    assert(current_level < number_of_levels_ - 1 && 0 <= current_level);
-    if (!interpolation_)
-        throw std::runtime_error("Interpolation not initialized.");
-
-    interpolation_->applyExtrapolatedRestriction(levels_[current_level].grid(), levels_[current_level + 1].grid(),
-                                                 result, x);
-}
-
-template <concepts::DomainGeometry DomainGeometry, concepts::DensityProfileCoefficients DensityProfileCoefficients>
-void GMGPolar<DomainGeometry, DensityProfileCoefficients>::FMGInterpolation(int current_level, Vector<double> result,
-                                                                            ConstVector<double> x) const
-{
-    assert(current_level < number_of_levels_ && 1 <= current_level);
-    if (!interpolation_)
-        throw std::runtime_error("Interpolation not initialized.");
-
-    interpolation_->applyFMGInterpolation(levels_[current_level].grid(), levels_[current_level - 1].grid(), result, x);
-}
