+++ src/CMakeLists.txt
@@ -19,8 +19,6 @@ set(POLAR_GRID_SOURCES
 # file(GLOB_RECURSE GMG_POLAR_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/GMGPolar/*.cpp)
 set(GMG_POLAR_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/GMGPolar/gmgpolar.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/GMGPolar/setup.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/GMGPolar/solver.cpp
 )
 
 # file(GLOB_RECURSE STENCIL_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/Stencil/*.cpp)
